+++ packages/permissionless/experimental/pimlico/index.ts
@@ -1,2 +1,2 @@
-import { prepareUserOperationForErc20Paymaster } from "./prepareUserOperationForErc20Paymaster"
+import { prepareUserOperationForErc20Paymaster } from "./utils/prepareUserOperationForErc20Paymaster"
 export { prepareUserOperationForErc20Paymaster }
